pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "FastWrite"
version = "1.2.1"
description = "A module for generating AI-based code documentation and data flow diagrams."
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}  # Or specify a license string, e.g., "MIT"
authors = [
    { name="R-G-KJSIT", email="r.giridhar@somaiya.edu" }
]
dependencies = [
    "groq",
    "google-genai",
    "requests",
    "nltk",
    "python-dotenv",
    "openai",
    "rouge_metric"
]

[project.urls]
Homepage = "https://fastwrite-py.vercel.app/"
Repository = "https://github.com/R-G-KJSIT/FastWrite"



